Fix bug caused by passing a function returning DataFrames into iter
iter(f, sentinel) calls f whenever _next_ is called, and raises
StopIterationError when f returns sentinel. Internally, this results
in f()==sentinel being computed, which causes problems when f() is
a DataFrame, since == raises an error whenever a DataFrame is
compared to another non-DataFrame object, or a DataFrame with different
dimensions.